Fix waveform visual overrun glitch on Safari
If starting playback and then seeking by clicking on the waveform during playback, the cursor would not stop at the end of the audio, but would overrun and create a huge orange rectangle. This was because the onended callback did not get invoked on Safari, and the keepAlive flag never got reset. Now reworked this to not rely on the keepAlive flag.